Specifies sequelize version exactly
This isn't the first time that some new sequelize beta broke something -
and that's okay, since these are beta versions.

However, I'm sick of initializing the project on a new machine and
getting lots of new errors. So we will use the latest working beta (13)
for now and upgrade later.