New test for server time mocking
So, apparently a utc month in javascript is zero based.
This led to problems when calculating the voteEndingTime,
because the month of a lunchbreak date in the database is not.

I fixed this in the previous commit (4829fbc), but I still needed
new tests to prevent regression.

The problem is, the tests were not wrong, the mocked server time was.
Calling testServer.restart([...], '25.06.2019') let the server
simulate to run at the 25.07(!).2019, that's why the tests did not
fail when the lunchbreak date was also one month in the future.

Long story short: I added a new test to ensure the server time is
mocked correctly.