Extract requesting user in singleton module
Previously all functionality around the requesting user was scattered
in multiple files. The middleware for initalization and token verification
was in util.js, the class and helper functions in /classes/user and
middleware functions had to use res.locals.user to access the user object.

This has now changed, everything is now bundled up as a singleton module
in /classes/user.