Fix global user state
Benioff: He kinda forgot that javascript works asynchronous...

No, for real: I'm now passing the data of the requesting user with
res.locals (again). The previous implementation would fail if two users
sent requests at the same time. There would be no guarantee that the
global user module contains the correct data, which is obviously a
fatal error.

Because the controllers are dependent on the requesting user module,
I am now injecting the user into the controllers with a new util middleware.
The controllers are then available in res.locals.controllers.